Insurance AI Assistant/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
faiss-cpu
pypdf
sentence-transformers
python-dotenv
python-multipart